Fix an overly-large malloc(). Thanks to clang-analyzer for spotting the mismatch.
git-svn-id: http://code.icecube.wisc.edu/svn/projects/photospline/trunk@131640 16731396-06f5-0310-8873-f7f720988828

+++ private/util/evalsplinefits.c
@@ -42,7 +42,7 @@ int main(int argc, char **argv) {
 		usage();
 
 	x = malloc(table.ndim*sizeof(double));
-	centers = malloc(table.ndim*sizeof(double));
+	centers = malloc(table.ndim*sizeof(int));
 	iterdim = -1;
 	for (i = 0; i < table.ndim; i++) { 
 		sscanf(argv[i+2],"%lf",&x[i]);
